requirements.txt
fastapi
uvicorn[standard]
pandas
reportlab
python-multipart
aiosmtplib
python-dotenv
jinja2
httpx
